BcAppModelをAppTableに変更 #90
