--- plugins/baser-core/src/Model/Table/AppModel.php
@@ -1,24 +1,0 @@
-<?php
-// TODO : コード確認要
-return;
-/**
- * baserCMS :  Based Website Development Project <https://basercms.net>
- * Copyright (c) baserCMS Users Community <https://basercms.net/community/>
- *
- * @copyright       Copyright (c) baserCMS Users Community
- * @link            https://basercms.net baserCMS Project
- * @package         Baser.Model
- * @since           baserCMS v 0.1.0
- * @license         https://basercms.net/license/index.html
- */
-
-App::uses('BcAppModel', 'Model');
-
-/**
- * Class AppModel
- *
- * @package Baser.Model
- */
-class AppModel extends BcAppModel
-{
-}
